requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-core>=2.18.0
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-ollama>=0.1.0
orjson>=3.10.0
python-dateutil>=2.9.0
apscheduler>=3.10.4
requests>=2.32.0
streamlit>=1.33.0
